+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+20.6.0
+------
+
 New Features
 ~~~~~~~~~~~~
 * util-stats: Add two new Java-friendly methods to `StatsReceiver` (`addGauge` and `provideGauge`)